+++ library/laboratory/src/main/java/io/mehow/laboratory/DefaultOptionFactory.kt
@@ -12,4 +12,15 @@ public interface DefaultOptionFactory {
    * **Warning** – returned default option must be of the same type as [feature]. Otherwise it will throw an exception.
    */
   public fun <T : Feature<T>> create(feature: T): Feature<*>?
+
+  /**
+   * Creates a new [DefaultOptionFactory] that will first look for a default value in this factory and then in the
+   * other factory.
+   */
+  @JvmDefault
+  public operator fun plus(factory: DefaultOptionFactory): DefaultOptionFactory = object : DefaultOptionFactory {
+    override fun <T : Feature<T>> create(
+      feature: T,
+    ) = this@DefaultOptionFactory.create(feature) ?: factory.create(feature)
+  }
 }